+++ ec/modules/helpers.py
@@ -24,6 +24,8 @@ def exit(exit_code=0):
   
   Could also be used to exit from the calling scripts in a thread safe manner.
   """
+  core.processExitHooks()
+  
   if state.isExitHooked and not hasattr(sys, 'exitfunc'): # the function is called from the exit hook
     sys.stderr.flush()
     sys.stdout.flush()