+++ scripts/examples/nut_shell.py
@@ -8,7 +8,7 @@ Notes
 * The decorators task, arg and group are available as __builtins__, when the script is launched using ec.
 * DocStrings of the tasks could be used by external modules (like sphinx). This is one of the key factors of developing ec, apart from it's predecessor Commander.
 """
-from ec.ec import task, arg, group, module, member, call, settings
+from ec.ec import task, arg, group, module, member, call, settings, exit_hook
 from ec.utils import get
 from ec.types import regex
 
@@ -62,6 +62,10 @@ class intro:
 # importing other modules
 import simple
 member(simple) # member is used to expose imported members as the children of the current module
+
+@exit_hook
+def clean_up(): # exit hooks are called when ec exits. There may be more than one hook.
+  print ':)'
 
 module(desc='A module to test decorator based configuration.') # module is an optional call, used to configure the group that wraps current module.