+++ HISTORY.rst
@@ -212,4 +212,5 @@ Log
   * 0700  utils.custom is now a CustomType (was a function). The change is made to allow dynamic configuration.
   * 1010  Improved the API docs.
   * 1425  Moved the test targets from tests/support to tests/targets.
+  * 1455  Introduced ec.exit_hook, a decorator that helps with adding **cleanup** functions.